2.1.0 release

+++ nics/main/wizard/__init__.py
@@ -1,10 +1,10 @@
 import os
 import shutil
+import sys
 
 from mykit.kit.utils import printer, print_screen
 
 from ..constants import TMPL_DOCS_DIR_PTH
-from .inspect import inspect
 from .workflow_writer import workflow_writer
 from .settings_writer import settings_writer
 
@@ -17,14 +17,32 @@ def run():
     CONTAINER_DIR_PTH = os.path.join(CWD, 'docs')
     SETTINGS_FILE_PTH = os.path.join(CWD, 'docs', 'settings.txt')
 
-    ## inspection
-    inspect(CONTAINER_DIR_PTH, WORKFLOW_FILE_PTH)
+    ## already_use is True if the user has used NICS before and only wants to reinitialize; otherwise, False.
+    already_use = False
+    if os.path.isfile(WORKFLOW_FILE_PTH) and os.path.isdir(CONTAINER_DIR_PTH):
+        already_use = True
+
+    if not already_use:
+        if os.path.isfile(WORKFLOW_FILE_PTH) or os.path.isdir(CONTAINER_DIR_PTH):
+            printer(f'ERROR: Workflow file or docs/ folder already exist.')
+            sys.exit(1)
 
     ## intro
     print_screen(
         'Welcome to NICS!\n'
         '================\n\n'
     )
+
+    if already_use:
+        print(
+            'Workflow file and docs/ folder are found, will rewrite these following files:\n'
+            f'- Workflow file {repr(WORKFLOW_FILE_PTH)}\n'
+            f'- Settings file {repr(SETTINGS_FILE_PTH)}'
+        )
+        usr = input('Continue?  (Y/n): ')
+        if usr not in ['', 'y', 'Y']:
+            printer(f'INFO: Exited.')
+            sys.exit(0)
 
     ## saving user's information
     author = input('Enter your name: ')
@@ -47,9 +65,10 @@ def run():
 
     workflow_writer(WORKFLOW_FILE_PTH, author, email, gh_repo, main_branch_name)
 
-    printer(f"INFO: Copying 'docs/' folder.")
-    shutil.copytree(TMPL_DOCS_DIR_PTH, CONTAINER_DIR_PTH)
-    printer(f'INFO: Done, {repr(CONTAINER_DIR_PTH)} is created.')
+    if not already_use:
+        printer(f"INFO: Copying 'docs/' folder.")
+        shutil.copytree(TMPL_DOCS_DIR_PTH, CONTAINER_DIR_PTH)
+        printer(f'INFO: Done, {repr(CONTAINER_DIR_PTH)} is created.')
 
     settings_writer(SETTINGS_FILE_PTH, author, email, gh_username, gh_repo, main_branch_name)
